ipn/ipnlocal: move last unconditional gvisor import, complete ts_omit_netstack support
Fixes #17283
