+++ feature/buildfeatures/feature_netstack_disabled.go
@@ -7,7 +7,7 @@
 
 package buildfeatures
 
-// HasNetstack is whether the binary was built with support for modular feature "gVisor netstack (userspace networking) support (TODO; not yet omittable)".
+// HasNetstack is whether the binary was built with support for modular feature "gVisor netstack (userspace networking) support".
 // Specifically, it's whether the binary was NOT built with the "ts_omit_netstack" build tag.
 // It's a const so it can be used for dead code elimination.
 const HasNetstack = false
